feat(docker): Rewrite Docker build script
This is a near complete rewrite of the Docker build script that is in
the process of being adopted across all of our CircleCI builds.

It includes the following changes:
- Moves building of the `latest` tag from `production` to `master`
branch
- Improve the management of the different tags we push to docker. It now
uses an array and so could in theory be extended however we wanted
- Remove the prefix `docker-` from any GitHub repositories when
constructing the Docker image name. I intend to use this to move, for
example `pelias/baseimage` to `pelias/docker-baseimage`

Connects https://github.com/pelias/pelias/issues/749