Support building on Github Actions or CircleCI
We recently moved the Pelias project away from TravisCI, and are now
running our unit tests on Github Actions.

Docker image builds, though, are still run on CircleCI. It might be nice
to move everything to a single CI provider, so this PR allows the Docker
image build script to work on _either_ CircleCI or Github Actions.